Implement basic OpenAI Responses API (#1)
This change adds a backend which allows using the OpenAI responses API.

Since Home Assistant does not expose any way to list or select
historical conversations, this implementation keeps an id and the date
it was created. If the next reply is within a user-configurable timeout,
new requests will use the old ID, meaning even if you close the chat
window, etc., you can continue the conversation.

The responses API might be more exciting in the future when/if Home
Assistant adds support for resuming conversation, which I heard on the
Discord might be in the roadmap.

![image](https://github.com/user-attachments/assets/bbc86c46-7a7d-4d51-98d1-7be56e922cc3)